Migrate to Typescript and split into modules
