+++ packages/command-tests/lib/errors.ts
@@ -1,0 +1,9 @@
+export function ensureError(input: unknown): Error {
+    if (input instanceof Error) {
+        return input;
+    }
+    if (typeof input === "string") {
+        return new Error(input);
+    }
+    return new Error(`Unknown Error: ${input}`);
+}
