+++ packages/command-tests/package.json
@@ -4,32 +4,31 @@
   "description": "Cli tests for widgets generator and tools",
   "private": true,
   "type": "module",
-  "main": "commands.js",
+  "main": "main.ts",
   "scripts": {
-    "start": "node commands.js"
+    "start": "node --experimental-strip-types main.ts",
+    "test": "tsc --noEmit"
   },
-  "devDependencies": {
+  "engines": {
+    "node": "^22.18.0"
+  },
+  "dependencies": {
     "@mendix/generator-widget": "workspace:*",
     "@mendix/pluggable-widgets-tools": "workspace:*",
-    "@prettier/plugin-xml": "^2.2.0",
-    "@types/jest-image-snapshot": "^4.3.1",
-    "@types/node": "^20.14.8",
-    "@types/shelljs": "^0.10.0",
-    "@types/xml2js": "^0.4.5",
     "async-mutex": "^0.5.0",
-    "chalk": "^5.6.2",
-    "eslint-config-prettier": "^8.5.0",
     "fs-extra": "^11.3.4",
-    "globals": "^17.6.0",
-    "identity-obj-proxy": "^3.0.0",
-    "peggy": "^1.2.0",
     "shelljs": "^0.10.0",
-    "shx": "^0.4.0",
     "tree-kill": "^1.2.2",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.8.4",
     "yeoman-environment": "^6.0.1",
     "yeoman-test": "^11.3.1"
+  },
+  "devDependencies": {
+    "@tsconfig/node-ts": "^23.6.4",
+    "@tsconfig/node22": "^22.0.5",
+    "@types/fs-extra": "^11.0.4",
+    "@types/node": "^20.14.8",
+    "@types/shelljs": "^0.10.0",
+    "typescript": "^5.9.3"
   },
   "keywords": [],
   "author": "",
