+++ packages/command-tests/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": [
+    "@tsconfig/node22/tsconfig.json",
+    "@tsconfig/node-ts/tsconfig.json"
+  ],
+  "compilerOptions": {
+"allowImportingTsExtensions": true
+  }
+}
